+++ src/test/php/lang/ast/unittest/emit/ArrayTypesTest.class.php
@@ -1,0 +1,27 @@
+<?php namespace lang\ast\unittest\emit;
+
+/**
+ * Array types
+ *
+ * @see  https://docs.hhvm.com/hack/types/type-system
+ */
+class ArrayTypesTest extends EmittingTest {
+
+  #[@test]
+  public function int_array_type() {
+    $t= $this->type('class <T> {
+      private array<int> $test;
+    }');
+
+    $this->assertEquals('int[]', $t->getField('test')->getType()->getName());
+  }
+
+  #[@test]
+  public function int_map_type() {
+    $t= $this->type('class <T> {
+      private array<string, int> $test;
+    }');
+
+    $this->assertEquals('[:int]', $t->getField('test')->getType()->getName());
+  }
+}